Reuse in-memory runtime graph and extract applyRuntimes
